--- frontend/src/components/pages/LoginPage.tsx
@@ -1,5 +1,0 @@
-import React, {createContext, useContext} from 'react';
-
-export function LoginPage() {
-  return <div>Login page</div>
-}
